requirements.txt
pydantic-ai
fastembed
qdrant-client
python-dotenv
redis
fastapi
uvicorn
slowapi
google-genai
